use Int

+++ src/commonMain/kotlin/io/github/tokuhirom/kdary/internal/BitVectorBuilder.kt
@@ -13,8 +13,7 @@ internal class BitVectorBuilder {
      * @param id the index of the bit.
      * @return the bit value at the specified index.
      */
-    operator fun get(id: UInt): Boolean =
-        (units[(id / BitVector.UNIT_SIZE.toUInt()).toInt()] shr (id % BitVector.UNIT_SIZE.toUInt()).toInt() and 1u) == 1u
+    operator fun get(id: Int): Boolean = (units[id / BitVector.UNIT_SIZE] shr id % BitVector.UNIT_SIZE and 1u) == 1u
 
     /**
      * Sets the bit at the specified index.
